@@ -43,18 +43,18 @@
   ],
   "description": "Get started with a Universal boilerplate containing tools like React, Redux, Koa...",
   "engines": {
-    "node": ">=6.0.0",
-    "npm": "^3.8.6"
+    "node": ">=7.6.0",
+    "npm": "^4.2.0"
   },
   "main": "src/server/index.js",
   "dependencies": {
     "autoprefixer": "^6.7.7",
-    "babel-core": "^6.24.0",
+    "babel-core": "^6.24.1",
     "babel-runtime": "^6.23.0",
-    "better-npm-run": "0.0.14",
-    "debug": "^2.6.3",
+    "better-npm-run": "0.0.15",
+    "debug": "^2.6.6",
     "extract-text-webpack-plugin": "^2.1.0",
-    "fs-extra": "^2.0.0",
+    "fs-extra": "^3.0.0",
     "globby": "^6.1.0",
     "inquirer": "^3.0.6",
     "ip": "^1.1.5",
@@ -65,73 +65,74 @@
     "koa-html-minifier": "^1.0.1",
     "koa-proxy": "^0.8.0",
     "koa-static": "^3.0.0",
-    "koa-static-cache": "^4.0.0",
-    "koa-webpack-dev-middleware": "^1.4.4",
+    "koa-static-cache": "^5.0.1",
+    "koa-webpack-dev-middleware": "^1.4.5",
     "koa-webpack-hot-middleware": "^1.0.3",
-    "normalize.css": "^5.0.0",
+    "normalize.css": "^6.0.0",
+    "prop-types": "^15.5.8",
     "purifycss-webpack-plugin": "^2.0.3",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
-    "react-helmet": "^4.0.0",
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
+    "react-helmet": "^5.0.3",
     "react-hot-loader": "^3.0.0-beta.6",
-    "react-redux": "^5.0.3",
+    "react-redux": "^5.0.4",
     "react-router": "^3.0.2",
     "react-router-redux": "^4.0.8",
     "redux": "^3.6.0",
     "serialize-javascript": "^1.3.0",
-    "source-map-support": "^0.4.12",
-    "webpack": "^2.2.1",
+    "source-map-support": "^0.4.15",
+    "webpack": "^2.4.1",
     "webpack-isomorphic-tools": "^3.0.2",
     "webpack-md5-hash": "^0.0.5",
     "why-did-you-update": "^0.0.8"
   },
   "devDependencies": {
-    "babel-cli": "^6.24.0",
-    "babel-eslint": "^7.1.0",
-    "babel-loader": "^6.4.0",
-    "babel-plugin-istanbul": "^4.0.0",
+    "babel-cli": "^6.24.1",
+    "babel-eslint": "^7.2.3",
+    "babel-loader": "^7.0.0",
+    "babel-plugin-istanbul": "^4.1.1",
     "babel-plugin-system-import-transformer": "^3.1.0",
     "babel-plugin-transform-react-jsx-source": "^6.22.0",
     "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-preset-latest": "^6.24.0",
-    "babel-preset-react": "^6.23.0",
+    "babel-preset-latest": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
     "babel-preset-react-optimize": "^1.0.1",
-    "babel-preset-stage-0": "^6.22.0",
+    "babel-preset-stage-0": "^6.24.1",
     "chai": "^3.5.0",
-    "clean-webpack-plugin": "^0.1.15",
-    "css-loader": "^0.27.3",
-    "enzyme": "^2.7.1",
-    "eslint": "^3.17.1",
+    "clean-webpack-plugin": "^0.1.16",
+    "css-loader": "^0.28.0",
+    "enzyme": "^2.8.2",
+    "eslint": "^3.19.0",
     "eslint-config-airbnb": "^14.1.0",
-    "eslint-loader": "^1.6.3",
+    "eslint-loader": "^1.7.1",
     "eslint-plugin-import": "^2.2.0",
     "eslint-plugin-jsx-a11y": "^4.0.0",
-    "eslint-plugin-react": "^6.10.0",
-    "file-loader": "^0.10.1",
-    "image-webpack-loader": "^3.2.0",
+    "eslint-plugin-react": "^6.10.3",
+    "file-loader": "^0.11.1",
+    "image-webpack-loader": "^3.3.0",
     "json-loader": "^0.5.4",
-    "karma": "^1.5.0",
+    "karma": "^1.6.0",
     "karma-chrome-launcher": "^2.0.0",
     "karma-coverage": "^1.1.1",
     "karma-mocha": "^1.3.0",
     "karma-phantomjs-launcher": "^1.0.4",
     "karma-sourcemap-loader": "^0.3.7",
-    "karma-spec-reporter": "^0.0.30",
-    "karma-webpack": "^2.0.2",
-    "mocha": "^3.2.0",
-    "node-sass": "^4.5.0",
+    "karma-spec-reporter": "^0.0.31",
+    "karma-webpack": "^2.0.3",
+    "mocha": "^3.3.0",
+    "node-sass": "^4.5.2",
     "nodemon": "^1.11.0",
     "npm-run-all": "^4.0.2",
-    "postcss": "^5.2.16",
-    "postcss-cssnext": "^2.9.0",
+    "postcss": "^5.2.17",
+    "postcss-cssnext": "^2.10.0",
     "postcss-loader": "^1.3.3",
-    "postcss-smart-import": "^0.6.10",
-    "react-addons-test-utils": "^15.4.2",
+    "postcss-smart-import": "^0.6.12",
+    "react-addons-test-utils": "^15.5.1",
     "react-hot-loader": "^3.0.0-beta.6",
     "sass-loader": "^6.0.3",
-    "sinon": "^1.17.7",
-    "sinon-chai": "^2.8.0",
-    "style-loader": "^0.13.2",
+    "sinon": "^2.1.0",
+    "sinon-chai": "^2.9.0",
+    "style-loader": "^0.16.1",
     "url-loader": "^0.5.8",
     "webpack-node-externals": "^1.5.4"
   },
