@@ -9,7 +9,7 @@
   "main": "server.js",
   "dependencies": {
     "babel-runtime": "^6.23.0",
-    "debug": "^2.6.3",
+    "debug": "^2.6.6",
     "ip": "^1.1.5",
     "koa": "^2.2.0",
     "koa-compress": "^2.0.0",
@@ -17,16 +17,16 @@
     "koa-helmet": "^3.1.0",
     "koa-html-minifier": "^1.0.1",
     "koa-static": "^3.0.0",
-    "koa-static-cache": "^4.0.0",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
-    "react-helmet": "^4.0.0",
-    "react-redux": "^5.0.3",
+    "koa-static-cache": "^5.0.1",
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
+    "react-helmet": "^5.0.3",
+    "react-redux": "^5.0.4",
     "react-router": "^3.0.2",
     "react-router-redux": "^4.0.8",
     "redux": "^3.6.0",
     "serialize-javascript": "^1.3.0",
-    "source-map-support": "^0.4.12"
+    "source-map-support": "^0.4.15"
   },
   "scripts": {
     "start": "DEBUG=app:* node server.js"
